+++ docs/release-notes.md
@@ -8,9 +8,19 @@ hide:
 
 ## 0.21.2
 
+### Added
+
+- PlaceholderField.
+- StrictModel which forbids extra attributes.
+
+### Changed
+
+- Validate on assignment is enabled.
+
 ### Fixed
 
-- Validators were defunct.
+- Pydantic validators are fixed. Field based ones as well as model based ones.
+- `__dict__` was wiped out when intializing a model.
 
 ## 0.21.1